app/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cloudinary==1.36.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==4.2.6
django-cloudinary-storage==0.3.0
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-filter==23.3
django-jazzmin==2.6.0
django-oauth-toolkit==2.3.0
django-paypal==2.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
shortuuid==1.0.11
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==7.1.0
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
wrapt==1.15.0
